Fix switch to Reserved Area forcing BackOffice IO logout (#1347)
Fix the switch to Reserved Area, forcing BackOffice IO logout.

This PR addresses a bug that occurred when a user switched from the
BackOffice IO to the Reserved Area, changed their institution, and then
returned to the BackOffice.
Previously, this sequence of actions could lead to the incorrect
institution being displayed.
This fix resolves the issue by forcing a BackOffice IO logout whenever a
user switches to the Reserved Area, ensuring that the correct entity is
always shown upon their return.

Closes IOPAE-2155 Resolves IOPAE-2156